+++ docs/source/news.rst
@@ -17,3 +17,6 @@ Features:
 * Added support for Python 3.3.
 * Actions can now be explicitly finished using a public API: ``Action.finish()``.
 * ``Action.context()`` context manager allows setting an action context without finishing the action when exiting the block.
+* Added a new API for Twisted ``Deferred`` support: ``eliot.twisted.DeferredContext``.
+  The older APIs (``Action.runCallback`` and ``Action.finishAfter``) are now deprecated.
+